Crate fs_set_times

fs-set-times provides functions to set timestamps on files, directories, and other filesystem objects.

On Windows, modifying a file’s timestamp requires write access to the file.

Enums

SystemTimeSpec

A value for specifying a time.

Traits

SetTimes

An extension trait for std::fs::File, cap_std::fs::File, and similar types.

Functions

set_atime

Set the last access timestamp of a file or other filesystem object.

set_mtime

Set the last modification timestamp of a file or other filesystem object.

set_symlink_times

Like set_times, but never follows symlinks.

set_times

Set the last access and last modification timestamps of a file or other filesystem object.